# public/swagger.local.json
{
    "swagger": "2.0",
    "info": {
        "title": "Donkeylift API",
        "description": "A web API to manage SQL databases (sqlite).",
        "version": "1.0.0-alpha"
    },
    "host": "localhost:3000",
    "schemes": [
        "http"
    ],
    "basePath": "/",
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/{account}/{database}": {
            "get": {
                "summary": "Get Database",
                "description": "Get database schema definition and a few other properties.",
                "operationId": "getInfo",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "description": "account that holds database",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "database",
                        "in": "path",
                        "description": "database to fetch",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "database"
                ],
                "responses": {
                    "200": {
                        "description": "Schema object",
                        "schema": {
                            "$ref": "#/definitions/Schema"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Create Database",
                "description": "Create or overwrite database with an empty database defined by schema payload. Returns an error if database exists and is not empty.",
                "operationId": "createDatabase",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "description": "account of database",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "database",
                        "in": "path",
                        "description": "database name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "schema",
                        "in": "body",
                        "description": "Database definition",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Schema"
                        }
                    }
                ],
                "tags": [
                    "database"
                ],
                "responses": {
                    "200": {
                        "description": "Database created."
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "patch": {
                "summary": "Patch Database",
                "description": "Modify Database applying a list of patches. Allows modifications of non-empty databases.",
                "operationId": "patchDatabase",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "description": "account holding database",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "database",
                        "in": "path",
                        "description": "database name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "patches",
                        "in": "body",
                        "description": "array with patches",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Patches"
                        }
                    }
                ],
                "tags": [
                    "database"
                ],
                "responses": {
                    "200": {
                        "description": "Patches applied."
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Database",
                "description": "Delete existing database.",
                "operationId": "delDatabase",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "description": "account holding database",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "database",
                        "in": "path",
                        "description": "database to fetch",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "force",
                        "in": "query",
                        "description": "Force delete database, set true to delete non-empty database.",
                        "required": false,
                        "type": "boolean"
                    }
                ],
                "tags": [
                    "database"
                ],
                "responses": {
                    "200": {
                        "description": "Database deleted."
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
        "/{account}/{database}/{table}.stats": {
            "get": {
                "summary": "Get row statistics",
                "operationId": "getStats",
                "description": "Returns min, max of result rows. See getRows for parameter description.",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "description": "database account",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "database",
                        "in": "path",
                        "description": "table database",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "table",
                        "in": "path",
                        "description": "table with data rows",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "$filter",
                        "in": "query",
                        "description": "array of row filters.",
                        "required": false,
                        "type": "array",
                        "collectionFormat": "tsv",
                        "items": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "$select",
                        "in": "query",
                        "description": "array of table fields.",
                        "required": false,
                        "type": "array",
                        "collectionFormat": "csv",
                        "items": {
                            "type": "string"
                        }
                    }
                ],
                "tags": [
                    "table"
                ],
                "responses": {
                    "200": {
                        "description": "A row result set"
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/{account}/{database}/{table}.objs": {
            "get": {
                "summary": "Get rows as objects",
                "operationId": "getObjs",
                "description": "Get rows from a table or a set of related tables and return them as JSON object trees. All method parameters from getRows are  supported.\n",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "description": "database account",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "database",
                        "in": "path",
                        "description": "table database",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "table",
                        "in": "path",
                        "description": "table with data rows",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "$filter",
                        "in": "query",
                        "description": "array of row filters.",
                        "required": false,
                        "type": "array",
                        "collectionFormat": "tsv",
                        "items": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "$select",
                        "in": "query",
                        "description": "array of table fields.",
                        "required": false,
                        "type": "array",
                        "collectionFormat": "csv",
                        "items": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "$orderby",
                        "in": "query",
                        "description": "array of order clauses.",
                        "required": false,
                        "type": "array",
                        "collectionFormat": "csv",
                        "items": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "$skip",
                        "in": "query",
                        "description": "skip this number of objects.",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "$top",
                        "in": "query",
                        "description": "max number of objects to be returned.",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "table"
                ],
                "responses": {
                    "200": {
                        "description": "An array of result objects.",
                        "schema": {
                            "$ref": "#/definitions/ResponseRows"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "User": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "owner",
                        "writer",
                        "reader"
                    ]
                }
            }
        },
        "Schema": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "tables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Table"
                    }
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/User"
                    }
                },
                "join_trees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/JoinTree"
                    }
                }
            }
        },
        "Table": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "row_alias": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "props": {
                    "type": "object"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Field"
                    }
                }
            }
        },
...
        "ResponseRows": {
            "type": "object",
            "properties": {
                "rows": {
                    "$ref": "#/definitions/Rows"
                },
                "count": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                },
                "nextUrl": {
                    "type": "string"
                }
            }
        },
...
        "Patches": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/Patch"
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}